Add a check for the Darwin dynamic linker. Use AC_TRY_LINK when checking
        * configure.in: Add a check for the Darwin dynamic linker. Use
        AC_TRY_LINK when checking for "nonposix getpwuid_r" so it notices
        "no getpwuid_r" correctly.

        * testglib.c (main): Make template[] bigger to prevent an overrun.
        Remove an unused variable. Initialize error to NULL.

        * tests/gio-test.c (main): Add a cast to prevent a warning when
        size_t is a long.

        * tests/type-test.c (main): Add an #ifdef to prevent a warning
        when G_HAVE_GINT64 is defined and G_GINT64_FORMAT isn't.

+++ tests/gio-test.c
@@ -399,7 +399,7 @@ main (int    argc,
 	  for (j = 0; j < buflen; j++)
 	    buf[j] = ' ' + ((buflen + j) % 95);
 	  g_print ("gio-test: child writing %d+%d bytes to %d\n",
-		   sizeof(i) + sizeof(buflen), buflen, writefd);
+		   (int)(sizeof(i) + sizeof(buflen)), buflen, writefd);
 	  write (writefd, &i, sizeof (i));
 	  write (writefd, &buflen, sizeof (buflen));
 	  write (writefd, buf, buflen);
